chore: refactor pages testing structure
