+++ .changeset/real-books-move.md
@@ -1,0 +1,5 @@
+---
+'@sigle/app': patch
+---
+
+Refactor pages testing structure.
