--- sigle/src/pages/_error.tsx
@@ -1,1 +1,0 @@
-export { MyError as default } from '../pages-lib/_error';
